allow float for duration_seconds field
yt-dlp returns duration as float (e.g. 35.966) for some videos.
Pydantic was rejecting it with int_from_float validation error.
Changed duration_seconds type from Optional[int] to Optional[float].

+++ backend/app/models/schemas.py
@@ -22,7 +22,7 @@ class VideoMetadata(BaseModel):
     engagement_rate: Optional[float] = None   # computed
     hashtags: List[str] = []
     upload_date: Optional[str] = None
-    duration_seconds: Optional[int] = None
+    duration_seconds: Optional[float] = None
     thumbnail_url: Optional[str] = None
